Letzte Kommentare geschrieben. BaseTest Klasse mit DAOTests zusammengeführt wegen SonarLint
