+++ bogenliga/bogenliga-business-logic/src/test/java/de/bogenliga/application/business/ligapasse/impl/dao/BaseLigapasseTest.java
@@ -1,20 +1,38 @@
-package de.bogenliga.application.business.ligapasse.impl;
+package de.bogenliga.application.business.ligapasse.impl.dao;
 
-import java.time.OffsetDateTime;
+import java.lang.reflect.InvocationTargetException;
+import java.util.Collections;
 import java.util.HashMap;
+import org.junit.Before;
 import org.junit.Rule;
+import org.junit.Test;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
+import de.bogenliga.application.business.baseClass.impl.BasicTest;
 import de.bogenliga.application.business.ligapasse.impl.entity.LigapasseBE;
+import de.bogenliga.application.common.component.dao.BasicDAO;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.Mockito.*;
 
 /**
- * TODO [AL] class documentation
- *
- * @author Andre Lehnert, eXXcellent solutions consulting & software gmbh
+ * @author Christopher Luzzi, Base for Ligapasse Tests & Tests for LigapasseDAO
  */
+
 public class BaseLigapasseTest {
     @Rule
     public MockitoRule mockitoRule = MockitoJUnit.rule();
+
+    @Mock
+    private BasicDAO basicDAO;
+
+    @InjectMocks
+    private LigapasseDAO underTest;
+
+    private LigapasseBE expectedBE;
+
+
 
     protected static final Long LIGAPASSE_WETTKAMPF_ID = 1L;
     protected static final Long LIGAPASSE_MATCH_ID = 2L;
@@ -66,5 +84,50 @@ public class BaseLigapasseTest {
 
     public HashMap<String, Object> getValuesToMethodMap(){return valuesToMethodMap; }
 
+
+    /**
+     * Tests for LigapasseDAO
+     */
+
+    //Implements generic way to test business entities methods
+    private BasicTest<LigapasseBE, LigapasseBE> basicDAOTest;
+
+    @Before
+    public void testSetup(){
+        expectedBE = getLigapasseBE();
+        basicDAOTest = new BasicTest<>(expectedBE, getValuesToMethodMap());
+        //configure mocks
+        when(basicDAO.selectSingleEntity(any(), any(), any())).thenReturn(expectedBE);
+
+        when(basicDAO.selectEntityList(any(), any(), any())).thenReturn(Collections.singletonList(expectedBE));
+    }
+
+
+    @Test
+    public void testAllFindMethods() throws InvocationTargetException, IllegalAccessException{
+        basicDAOTest.testAllFindMethods(underTest);
+    }
+
+    @Test
+    public void findLigapassenByLigamatchId(){
+        try{
+            basicDAOTest.testAllFieldsOnEqualToExpectedEntity(underTest.findLigapasseByLigamatchId(LIGAPASSE_MATCH_ID));
+        } catch (InvocationTargetException e) {
+            e.printStackTrace();
+        } catch (IllegalAccessException e) {
+            e.printStackTrace();
+        }
+    }
+
+    @Test
+    public void findLigapasseByLigamatchId(){
+        try{
+            basicDAOTest.testAllFieldsOnEqualToExpectedEntity(underTest.findLigapasseByLigamatchId(LIGAPASSE_MATCH_ID));
+        } catch (InvocationTargetException e) {
+            e.printStackTrace();
+        } catch (IllegalAccessException e) {
+            e.printStackTrace();
+        }
+    }
 
 }
