+++ bogenliga/bogenliga-business-logic/src/test/java/de/bogenliga/application/business/passe/impl/business/PasseComponentImplTest.java
@@ -11,11 +11,10 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
-import de.bogenliga.application.business.ligapasse.impl.BaseLigapasseTest;
+import de.bogenliga.application.business.ligapasse.impl.dao.BaseLigapasseTest;
 import de.bogenliga.application.business.ligapasse.impl.dao.LigapasseDAO;
 import de.bogenliga.application.business.ligapasse.impl.entity.LigapasseBE;
 import de.bogenliga.application.business.ligapasse.impl.mapper.LigapasseToPasseMapper;
-import de.bogenliga.application.business.match.api.types.MatchDO;
 import de.bogenliga.application.business.passe.api.types.PasseDO;
 import de.bogenliga.application.business.passe.impl.dao.PasseBaseDAOTest;
 import de.bogenliga.application.business.passe.impl.dao.PasseDAO;
